Move GetDepositSnapshot to beacon APIs package (cause official API already)
